Test retrieval of genre position map from the cache > > Co-authored-by: Iliana Scalco <dokratos@hotmail.it>
